remove images close or on top of stars in the stereo reconstruction; display stars in evndisp
git-svn-id: https://svnsrv.desy.de/desy/EVNDISP/trunk@1717 766874ec-0625-4018-90f7-740f284a29e6

+++ src/VReadRunParameter.cpp
@@ -266,7 +266,7 @@ bool VReadRunParameter::readCommandline( int argc, char *argv[] )
         }
 	else if( iTemp.find( "starbrightness" ) < iTemp.size() )
 	{
-	    fRunPara->fMinStarBrightness_B = atoi( iTemp.substr( iTemp.rfind( "=" )+1, iTemp.size() ).c_str() );
+	    fRunPara->fMinStarBrightness_B = atof( iTemp.substr( iTemp.rfind( "=" )+1, iTemp.size() ).c_str() );
         }
 // write all pixel to dst file
         else if( iTemp.find( "dstallpixel" ) < iTemp.size() )
